+++ staticinline/apps.py
@@ -1,5 +1,6 @@
-from base64 import b64encode
 import mimetypes
+import base64
+import hashlib
 
 from django.apps import AppConfig
 
@@ -21,21 +22,24 @@ class StaticInlineAppConfig(AppConfig):
         return {
             'base64': self.encode_base64,
             'data': self.encode_data_uri,
+            'sri': self.encode_sri,
         }
 
     def encode_base64(self, data, path):
         """
         Encodes with standard Base64.
+
         :param bytes data: Input data to encode.
         :return: Base64 encoded input string.
         :rtype: str
         :raises Exception: if the file is not suitable to be Base64 encoded.
         """
-        return b64encode(data).decode(self.encoder_response_format)
+        return base64.b64encode(data).decode(self.encoder_response_format)
 
     def encode_data_uri(self, data, path):
         """
         Convert to the data URI scheme.
+
         :param bytes data: Input data to encode.
         :return: data URI string.
         :rtype: str
@@ -46,3 +50,28 @@ class StaticInlineAppConfig(AppConfig):
         else:
             prefix = 'data:{0};base64,'.format(mimetype)
         return prefix + self.encode_base64(data, data)
+
+    def encode_sri(self, data, path):
+        """
+        Adds Subresource Integrity encoder to staticinline. Read more:
+
+        https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity
+
+        Example Usage::
+
+          {% load staticfiles %}
+          {% load staticinline %}
+
+          <link
+            rel="stylesheet"
+            href="{% static "base.css" %}"
+            integrity="{% staticinline "base.css" encode="sri" %}"
+            crossorigin="anonymous"/>
+
+        :param bytes data: File content to build the SRI hash of.
+        :return: sha256 hash
+        :rtype: str
+        """
+        hash = hashlib.sha256(data).digest()
+        hash_base64 = base64.b64encode(hash).decode()
+        return 'sha256-{}'.format(hash_base64)