+++ CHANGELOG.rst
@@ -7,12 +7,14 @@ Changelog
 ========
 
 - Added support for Django 2.1 and Python 3.7.
+- Added proper documentation.
+- Added ``sri`` (Subresource Integrity) encoder to generate a sha256 for a given file.
 
 v1.1 (2018-08-09)
 =================
 
-- Support for custom data encoders to modify file content on the fly.
-- Adds ``data`` and ``base64`` encoders, both convert data into base64.
+- Added support for custom data encoders to modify file content on the fly.
+- Added ``data`` and ``base64`` encoders, both convert data into base64.
 
 v1.0 (2018-04-29)
 =================